geoconnect/requirements.txt
python-dotenv==0.8.2
pandas==0.19.2
numpy==1.11.0
requests==2.12.4